fix: Update example & redo tests
