+++ packages/tinkoff_acquiring/test/sdk_test_constant.dart
@@ -17,18 +17,19 @@ const String publicKey =
     'MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA5Yg3RyEkszggDVMDHCAGzJm0mYpYT53BpasrsKdby8iaWJVACj8ueR0Wj3Tu2BY64HdIoZFvG0v7UqSFztE/zUvnznbXVYguaUcnRdwao9gLUQO2I/097SHF9r++BYI0t6EtbbcWbfi755A1EWfu9tdZYXTrwkqgU9ok2UIZCPZ4evVDEzDCKH6ArphVc4+iKFrzdwbFBmPmwi5Xd6CB9Na2kRoPYBHePGzGgYmtKgKMNs+6rdv5v9VB3k7CS/lSIH4p74/OPRjyryo6Q7NbL+evz0+s60Qz5gbBRGfqCA57lUiB3hfXQZq5/q1YkABOHf9cR6Ov5nTRSOnjORgPjwIDAQAB';
 
 final String cardDataNo3DS = CardData(
-  pan: '4300000000000777',
-  expDate: '1122',
-  cvv: '111',
+  pan: '2201382000000005',
+  expDate: '1224',
+  cvv: '123',
   cardHolder: 'T. TESTING',
 ).encode(publicKey);
 
 // For mobile, use password: 1234
 // For web, use password: password
+// ACS 1qwezxc
 final String cardData3DS = CardData(
-  pan: '5411420000000002',
-  expDate: '1122',
-  cvv: '111',
+  pan: '2201382000000047',
+  expDate: '1224',
+  cvv: '123',
   cardHolder: 'T. TESTING',
 ).encode(publicKey);
